Rename ExperimentalResults -> ExperimentalRunResults and add a new ExperimentalResults to handle additional aggregation and plotting
