+++ tests/integration/models/conv/test_multisensory_classifier.py
@@ -1,4 +1,3 @@
-import os
 import tempfile
 import unittest
